+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 - Made a new `BlankCallback` type for timer
 - Changed timer code to use the `BlankCallback` in the CallbackStore.
+- Optimized Cargo.toml file to only include one dependency: piston.
 
 ## 0.1.14